+++ backend/internal/jwt/types.go
@@ -8,14 +8,14 @@ import (
 // AccessTokenClaims represents the claims for an access token.
 type AccessTokenClaims struct {
 	Role       db.UserRole `json:"role"`
-	UserID     int32       `json:"userId"`
-	OfficeID   *int32      `json:"officeId,omitempty"`
-	AdminRefID *int32      `json:"adminRefId,omitempty"`
+	UserID     int64       `json:"userId"`
+	OfficeID   *int64      `json:"officeId,omitempty"`
+	AdminRefID *int64      `json:"adminRefId,omitempty"`
 	jwt.RegisteredClaims
 }
 
 // RefreshTokenClaims represents the claims for a refresh token.
 type RefreshTokenClaims struct {
-	UserID int32 `json:"userId"`
+	UserID int64 `json:"userId"`
 	jwt.RegisteredClaims
 }
